pangowin32-fontmap: fix wrong usage of GWeakRef in PangoWin32Font
Casting a `PangoFontMap *fontmap` field to a `GWeakRef` is not correct.

Either use GWeakRef API throughout, or don't.

In particular, assigning a plain pointer to the field (as done in
_pango_cairo_win32_font_new()) and later treat it as GWeakRef leads to a
crash.

Don't use `GWeakRef` API. g_object_add_weak_pointer() is sufficient.
It's not thread safe, but neither was the previous code.

Fixes: 16e4dd37e5fd ('Port PangoWin32 to use GWeakRef')